fix update-catalog.mjs move it to utils and simplify the code

+++ api/delete-catalogs.mjs
@@ -2,5 +2,9 @@ import fs from 'fs'
 import path from 'path'
 
 const catalogsPath = new URL('../src/data/catalogs.json', import.meta.url)
-fs.rmSync(catalogsPath)
-console.log(`Catalogs file deleted at ${path.resolve(catalogsPath.pathname)}`)
+if (fs.existsSync(catalogsPath)) {
+	fs.rmSync(catalogsPath)
+	console.log(`Catalogs file deleted at ${path.resolve(catalogsPath.pathname)}`)
+} else {
+	console.log(`Catalogs file does not exist at ${path.resolve(catalogsPath.pathname)}`)
+}
